Move notifications into separate plugin
