+++ packages/acter_notifify/.flutter-plugins
@@ -1,0 +1,27 @@
+# This is a generated file; do not edit or check into version control.
+acter_flutter_sdk=/DEV/acter/a3/packages/rust_sdk/
+device_info_plus=/home/ben/.pub-cache/hosted/pub.dev/device_info_plus-10.1.1/
+firebase_core=/home/ben/.pub-cache/hosted/pub.dev/firebase_core-3.3.0/
+firebase_core_web=/home/ben/.pub-cache/hosted/pub.dev/firebase_core_web-2.17.4/
+flutter_local_notifications=/home/ben/.pub-cache/hosted/pub.dev/flutter_local_notifications-17.2.1+2/
+flutter_local_notifications_linux=/home/ben/.pub-cache/hosted/pub.dev/flutter_local_notifications_linux-4.0.1/
+flutter_secure_storage=/home/ben/.pub-cache/hosted/pub.dev/flutter_secure_storage-9.2.2/
+flutter_secure_storage_linux=/home/ben/.pub-cache/hosted/pub.dev/flutter_secure_storage_linux-1.2.1/
+flutter_secure_storage_macos=/home/ben/.pub-cache/hosted/pub.dev/flutter_secure_storage_macos-3.1.2/
+flutter_secure_storage_web=/home/ben/.pub-cache/hosted/pub.dev/flutter_secure_storage_web-1.2.1/
+flutter_secure_storage_windows=/home/ben/.pub-cache/hosted/pub.dev/flutter_secure_storage_windows-3.1.2/
+package_info=/home/ben/.pub-cache/hosted/pub.dev/package_info-2.0.2/
+path_provider=/home/ben/.pub-cache/hosted/pub.dev/path_provider-2.1.4/
+path_provider_android=/home/ben/.pub-cache/hosted/pub.dev/path_provider_android-2.2.9/
+path_provider_foundation=/home/ben/.pub-cache/hosted/pub.dev/path_provider_foundation-2.4.0/
+path_provider_linux=/home/ben/.pub-cache/hosted/pub.dev/path_provider_linux-2.2.1/
+path_provider_windows=/home/ben/.pub-cache/hosted/pub.dev/path_provider_windows-2.3.0/
+push=/home/ben/.pub-cache/hosted/pub.dev/push-2.3.0/
+push_android=/home/ben/.pub-cache/hosted/pub.dev/push_android-0.6.0/
+push_ios=/home/ben/.pub-cache/hosted/pub.dev/push_ios-0.5.1/
+shared_preferences=/home/ben/.pub-cache/hosted/pub.dev/shared_preferences-2.3.1/
+shared_preferences_android=/home/ben/.pub-cache/hosted/pub.dev/shared_preferences_android-2.3.0/
+shared_preferences_foundation=/home/ben/.pub-cache/hosted/pub.dev/shared_preferences_foundation-2.5.0/
+shared_preferences_linux=/home/ben/.pub-cache/hosted/pub.dev/shared_preferences_linux-2.4.0/
+shared_preferences_web=/home/ben/.pub-cache/hosted/pub.dev/shared_preferences_web-2.4.1/
+shared_preferences_windows=/home/ben/.pub-cache/hosted/pub.dev/shared_preferences_windows-2.4.0/
