+++ packages/acter_notifify/lib/android.dart
@@ -1,12 +1,14 @@
 import 'dart:async';
 
-import 'package:acter/common/notifications/notifications.dart';
-import 'package:acter/common/notifications/util.dart';
+import 'package:acter_notifify/local.dart';
+import 'package:acter_notifify/util.dart';
+
+import 'notifications.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:logging/logging.dart';
 
-final _log = Logger('a3::notifications::android');
+final _log = Logger('a3::notifify::android');
 
 final Map<String, List<Message>> pendingMessage = {};
