+++ packages/acter_notifify/pubspec.yaml
@@ -11,8 +11,15 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  rust_sdk: 
+  acter_flutter_sdk: 
     path: ../rust_sdk
+
+  logging: ^1.2.0
+  flutter_local_notifications: ^17.1.2
+  device_info_plus: ^10.1.0
+  push: ^2.3.0
+  convert: ^3.1.1
+  firebase_core: ^3.1.0
 
 dev_dependencies:
   flutter_test:
