+++ packages/acter_notifify/test/acter_notifify_test.dart
@@ -3,10 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:acter_notifify/acter_notifify.dart';
 
 void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
+  test('import smoketest', () {
+    // doesn't do anything
   });
 }
